feat: 完成知识库图一跳ACL与实体词条即时回放功能开发
1. 新增知识可见性过滤逻辑，图一跳召回时隐藏私密邻居内容
2. 实体词条写入后立即调用SetWriteBackReplay重建索引
3. 重构工具参数归一逻辑，统一入口为NormalizeInvocation
4. 移除冗余的filenorm与hostexecnorm包装，简化工具链
5. 修复知识链接查询的ACL过滤条件
6. 新增配套测试用例与评测报告

+++ internal/data/monitor_alert.go
@@ -24,11 +24,13 @@ func (r *monitorRepo) ensureMonitorAlertFiringStateCols(ctx context.Context) {
 			col string
 			ddl string
 		}{
-			{"last_fired_at", `ALTER TABLE monitor_alert_rules ADD COLUMN last_fired_at INTEGER`},
+			// 毫秒时间戳列必须 BIGINT：读径按 UnixMilli 解释，PG INTEGER（int4）
+			// 放不下 ms epoch（≈1.7e12）→ 写入 22003 越界（2026-08-17 真机事故）。
+			{"last_fired_at", `ALTER TABLE monitor_alert_rules ADD COLUMN last_fired_at BIGINT`},
 			{"last_fired_value", `ALTER TABLE monitor_alert_rules ADD COLUMN last_fired_value REAL NOT NULL DEFAULT 0`},
-			{"last_fired_window_start", `ALTER TABLE monitor_alert_rules ADD COLUMN last_fired_window_start INTEGER`},
+			{"last_fired_window_start", `ALTER TABLE monitor_alert_rules ADD COLUMN last_fired_window_start BIGINT`},
 			{"firing_state", `ALTER TABLE monitor_alert_rules ADD COLUMN firing_state TEXT NOT NULL DEFAULT 'idle'`},
-			{"recovered_at", `ALTER TABLE monitor_alert_rules ADD COLUMN recovered_at INTEGER`},
+			{"recovered_at", `ALTER TABLE monitor_alert_rules ADD COLUMN recovered_at BIGINT`},
 		}
 		for _, a := range alters {
 			has, err := columnExistsWithDialect(ctx, db, "monitor_alert_rules", a.col, d)
